requirements.txt
# CrewAI Framework
crewai>=0.51.0
crewai-tools>=0.8.0

# Browser Automation
playwright>=1.40.0

# Data Processing
pandas>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
